// packages/runner-safety-web/src/protos/pipeline.ts
export interface Summary {
  /**
   * @generated from protobuf field: string cwd = 1;
   */
  cwd: string;
  /**
   * @generated from protobuf field: repeated safety_web.Violation safety_web_violations = 2;
   */
  safetyWebViolations: Violation[];
  /**
   * @generated from protobuf field: repeated safety_web.Violation safety_web_silenced_violations = 3;
   */
  safetyWebSilencedViolations: Violation[];
  /**
   * ESLint violations not related to the safety-web plugin
   *
   * @generated from protobuf field: repeated safety_web.Violation other_violations = 4;
   */
  otherViolations: Violation[];
  /**
   * @generated from protobuf field: repeated safety_web.Violation other_silenced_violations = 5;
   */
  otherSilencedViolations: Violation[];
}
/**
 * @generated from protobuf message safety_web.Package
 */
export interface Package {
  /**
   * @generated from protobuf field: string name = 1;
   */
  name: string;
  /**
   * @generated from protobuf field: string relative_path = 2;
   */
  relativePath: string;
  /**
   * @generated from protobuf field: string version = 3;
   */
  version: string;
  /**
   * @generated from protobuf field: safety_web.Summary safety_web_summary = 4;
   */
  safetyWebSummary?: Summary;
}
/**
 * @generated from protobuf message safety_web.PackageManager
 */
export interface PackageManager {
  /**
   * @generated from protobuf field: string kind = 1;
   */
  kind: string;
  /**
   * @generated from protobuf field: string version = 2;
   */
  version: string;
}
/**
 * An invocation of the pipeline on 1 repository. Captures characteristics found
 * while analyzing and installing the repository.
 *
 * @generated from protobuf message safety_web.Repository
 */
export interface Repository {
  /**
   * @generated from protobuf field: string url = 1;
   */
  url: string;
  /**
   * @generated from protobuf field: safety_web.PackageManager package_manager_found = 3;
   */
  packageManagerFound?: PackageManager;
  /**
   * @generated from protobuf field: safety_web.PackageManager package_manager_used = 4;
   */
  packageManagerUsed?: PackageManager;
  /**
   * @generated from protobuf field: repeated safety_web.Package packages = 5;
   */
  packages: Package[];
  /**
   * @generated from protobuf field: string logs = 6;
   */
  logs: string;
}
// ...
export const Summary = new Summary$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Package$Type extends MessageType<Package> {
  constructor() {
    super('safety_web.Package', [
      {no: 1, name: 'name', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
      {
        no: 2,
        name: 'relative_path',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/,
      },
      {no: 3, name: 'version', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
      {no: 4, name: 'safety_web_summary', kind: 'message', T: () => Summary},
    ]);
  }
  create(value?: PartialMessage<Package>): Package {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.name = '';
    message.relativePath = '';
    message.version = '';
    if (value !== undefined)
      reflectionMergePartial<Package>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: Package,
  ): Package {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string name */ 1:
          message.name = reader.string();
          break;
        case /* string relative_path */ 2:
          message.relativePath = reader.string();
          break;
        case /* string version */ 3:
          message.version = reader.string();
          break;
        case /* safety_web.Summary safety_web_summary */ 4:
          message.safetyWebSummary = Summary.internalBinaryRead(
            reader,
            reader.uint32(),
            options,
            message.safetyWebSummary,
          );
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: Package,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string name = 1; */
    if (message.name !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.name);
    /* string relative_path = 2; */
    if (message.relativePath !== '')
      writer.tag(2, WireType.LengthDelimited).string(message.relativePath);
    /* string version = 3; */
    if (message.version !== '')
      writer.tag(3, WireType.LengthDelimited).string(message.version);
    /* safety_web.Summary safety_web_summary = 4; */
    if (message.safetyWebSummary)
      Summary.internalBinaryWrite(
        message.safetyWebSummary,
        writer.tag(4, WireType.LengthDelimited).fork(),
        options,
      ).join();
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message safety_web.Package
 */
export const Package = new Package$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PackageManager$Type extends MessageType<PackageManager> {
  constructor() {
    super('safety_web.PackageManager', [
      {no: 1, name: 'kind', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
      {no: 2, name: 'version', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
    ]);
  }
  create(value?: PartialMessage<PackageManager>): PackageManager {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.kind = '';
    message.version = '';
    if (value !== undefined)
      reflectionMergePartial<PackageManager>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: PackageManager,
  ): PackageManager {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string kind */ 1:
          message.kind = reader.string();
          break;
        case /* string version */ 2:
          message.version = reader.string();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: PackageManager,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string kind = 1; */
    if (message.kind !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.kind);
    /* string version = 2; */
    if (message.version !== '')
      writer.tag(2, WireType.LengthDelimited).string(message.version);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message safety_web.PackageManager
 */
export const PackageManager = new PackageManager$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Repository$Type extends MessageType<Repository> {
  constructor() {
    super('safety_web.Repository', [
      {no: 1, name: 'url', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
      {
        no: 3,
        name: 'package_manager_found',
        kind: 'message',
        T: () => PackageManager,
      },
      {
        no: 4,
        name: 'package_manager_used',
        kind: 'message',
        T: () => PackageManager,
      },
      {
        no: 5,
        name: 'packages',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => Package,
      },
      {no: 6, name: 'logs', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
    ]);
  }
  create(value?: PartialMessage<Repository>): Repository {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.url = '';
    message.packages = [];
    message.logs = '';
    if (value !== undefined)
      reflectionMergePartial<Repository>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: Repository,
  ): Repository {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string url */ 1:
          message.url = reader.string();
          break;
        case /* safety_web.PackageManager package_manager_found */ 3:
          message.packageManagerFound = PackageManager.internalBinaryRead(
            reader,
            reader.uint32(),
            options,
            message.packageManagerFound,
          );
          break;
        case /* safety_web.PackageManager package_manager_used */ 4:
          message.packageManagerUsed = PackageManager.internalBinaryRead(
            reader,
            reader.uint32(),
            options,
            message.packageManagerUsed,
          );
          break;
        case /* repeated safety_web.Package packages */ 5:
          message.packages.push(
            Package.internalBinaryRead(reader, reader.uint32(), options),
          );
          break;
        case /* string logs */ 6:
          message.logs = reader.string();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: Repository,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string url = 1; */
    if (message.url !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.url);
    /* safety_web.PackageManager package_manager_found = 3; */
    if (message.packageManagerFound)
      PackageManager.internalBinaryWrite(
        message.packageManagerFound,
        writer.tag(3, WireType.LengthDelimited).fork(),
        options,
      ).join();
    /* safety_web.PackageManager package_manager_used = 4; */
    if (message.packageManagerUsed)
      PackageManager.internalBinaryWrite(
        message.packageManagerUsed,
        writer.tag(4, WireType.LengthDelimited).fork(),
        options,
      ).join();
    /* repeated safety_web.Package packages = 5; */
    for (let i = 0; i < message.packages.length; i++)
      Package.internalBinaryWrite(
        message.packages[i],
        writer.tag(5, WireType.LengthDelimited).fork(),
        options,
      ).join();
    /* string logs = 6; */
    if (message.logs !== '')
      writer.tag(6, WireType.LengthDelimited).string(message.logs);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message safety_web.Repository
 */
export const Repository = new Repository$Type();